[#154951783] Artist wants Chord position to have resolution finer than 1 beat, in order to accurately portray styles of music.  \   * [core/migration] V56 modify all chord position columns to FLOAT
[#154976066] Architect wants to limit the floating point precision of chord and event position, in order to limit obsession over the position of things.
 \
  * [core/model] Entities use shared roundPosition()

+++ core/src/main/java/io/xj/core/model/chord/Chord.java
@@ -20,7 +20,7 @@ public abstract class Chord extends Entity {
   static final String MARKER_NON_CHORD = "---";
 
   protected String name;
-  protected Integer position;
+  protected Double position;
 
   public String getName() {
     return name;
@@ -28,11 +28,11 @@ public abstract class Chord extends Entity {
 
   public abstract Chord setName(String name);
 
-  public Integer getPosition() {
+  public Double getPosition() {
     return position;
   }
 
-  public abstract Chord setPosition(Integer position);
+  public abstract Chord setPosition(Double position);
 
   public abstract BigInteger getParentId();
