+++ craft/src/test/java/io/xj/craft/macro/CraftFoundationNextMainIT.java
@@ -183,11 +183,11 @@ public class CraftFoundationNextMainIT {
     Iterator<LinkChord> it = resultLinkChords.iterator();
 
     LinkChord chordOne = it.next();
-    assertEquals(Integer.valueOf(0), chordOne.getPosition());
+    assertEquals(Double.valueOf(0), chordOne.getPosition());
     assertEquals("G minor", chordOne.getName());
 
     LinkChord chordTwo = it.next();
-    assertEquals(Integer.valueOf(8), chordTwo.getPosition());
+    assertEquals(Double.valueOf(8), chordTwo.getPosition());
     assertEquals("Ab minor", chordTwo.getName());
 
     // choice of macro-type pattern
